pyproject.toml
[tool.poetry]
name = "control-clinic"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
autopep8 = "^2.0.4"
flask = "^3.0.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
flask-wtf = "^1.2.1"
email-validator = "^2.1.0.post1"
flask-admin = "^1.6.1"
flask-bootstrap = "^3.3.7.1"
passlib = "^1.7.4"
flask-login = "^0.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
